Added min-/max values

+++ Gold.HealthTracker.DBModel/Set.cs
@@ -5,6 +5,6 @@ public class Set
     public int Id { get; set; }
     public int ExerciseId { get; set; }
     public virtual Exercise? Exercise { get; set; }
-    public int Repetitions { get; set; }
+    public float Repetitions { get; set; }
     public float Weight { get; set; }
 }
